+++ packages/doctor/src/Scanner.ts
@@ -1,6 +1,7 @@
 import * as path from 'path';
 import Timer from './Timer';
-import { IScannerOptions, IScanOptions, IFileInfo, IScannerReports } from './types/Scanner';
+import { IScannerOptions, IScanOptions, IScannerReports } from './types/Scanner';
+import { IFileInfo } from './types/File';
 import getCustomESLintConfig from './getCustomESLintConfig';
 import getEslintReports from './getEslintReports';
 import getMaintainabilityReports from './getMaintainabilityReports';
@@ -20,13 +21,7 @@ export default class Scanner {
     const timer = new Timer(options?.timeout);
     const reports = {} as IScannerReports;
 
-    const files: IFileInfo[] = getFiles(directory, this.options.supportExts, this.options.ignore);
-
-    // Process package.json file.
-    const packageFileInfo = getFiles(path.join(directory, 'package.json'), ['json'])[0];
-    if (packageFileInfo) {
-      files.push(packageFileInfo);
-    }
+    const files: IFileInfo[] = getFiles(directory, this.options.ignore);
 
     // Set files info
     reports.filesInfo = {
@@ -45,7 +40,7 @@ export default class Scanner {
         }
         customConfig.parserOptions.project = `${path.join(directory, './')}**/tsconfig.json`;
       }
-      reports.ESLint = getEslintReports(timer, files, ruleKey, customConfig, options?.fix);
+      reports.ESLint = getEslintReports(directory, timer, files, ruleKey, customConfig, options?.fix);
     }
 
     // Calculate maintainability
@@ -55,7 +50,7 @@ export default class Scanner {
 
     // Calculate repeatability
     if (!options || options.disableRepeatability !== true) {
-      reports.repeatability = await getRepeatabilityReports(directory, this.options.supportExts, this.options.ignore, options?.tempFileDir);
+      reports.repeatability = await getRepeatabilityReports(directory, this.options.ignore, options?.tempFileDir);
     }
 
     // Calculate total score