+++ packages/doctor/src/types/File.ts
@@ -1,0 +1,6 @@
+export interface IFileInfo {
+  path: string;
+  source: string;
+  // lines of code
+  LoC: number;
+}